group node, element and tree tests
